requirements.txt
aiohttp
asyncio
discord.py
pillow
speedtest-cli
animec